Rename `TradeOptionSelectScene`->`RampOptionSelectScene`
Same rationale as previous rename
