+++ src/types/routerTypes.tsx
@@ -50,6 +50,7 @@ import type { MigrateWalletCompletionParams } from '../components/scenes/Migrate
 import type { MigrateWalletSelectCryptoParams } from '../components/scenes/MigrateWalletSelectCryptoScene'
 import type { OtpRepairParams } from '../components/scenes/OtpRepairScene'
 import type { RampCreateParams } from '../components/scenes/RampCreateScene'
+import type { RampSelectOptionParams } from '../components/scenes/RampOptionSelectScene'
 import type { RequestParams } from '../components/scenes/RequestScene'
 import type { SendScene2Params } from '../components/scenes/SendScene2'
 import type { EarnSceneParams } from '../components/scenes/Staking/EarnScene'
@@ -64,7 +65,6 @@ import type { SweepPrivateKeyCalculateFeeParams } from '../components/scenes/Swe
 import type { SweepPrivateKeyCompletionParams } from '../components/scenes/SweepPrivateKeyCompletionScene'
 import type { SweepPrivateKeyProcessingParams } from '../components/scenes/SweepPrivateKeyProcessingScene'
 import type { SweepPrivateKeySelectCryptoParams } from '../components/scenes/SweepPrivateKeySelectCryptoScene'
-import type { RampSelectOptionParams } from '../components/scenes/TradeOptionSelectScene'
 import type { TransactionDetailsParams } from '../components/scenes/TransactionDetailsScene'
 import type { TransactionListParams } from '../components/scenes/TransactionListScene'
 import type { TransactionsExportParams } from '../components/scenes/TransactionsExportScene'
